fix: logout

+++ app/Http/Controllers/AuthController.php
@@ -80,13 +80,12 @@ class AuthController extends Controller
 
     public function logout(Request $request)
     {
-        $request->user()->currentAccessToken()->delete();
-
-        return response()->json(
-            [
-                'message' => 'Logged out successfully'
-            ],
-            200
-        );
+        if (auth()->check()) {
+            $request->user()->currentAccessToken()->delete();
+        }
+    
+        return response()->json([
+            'message' => 'Logged out successfully'
+        ], 200);
     }
 }
